backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.3
qdrant-client==1.12.1
psycopg2-binary==2.9.10
pydantic==2.10.6
python-multipart==0.0.20
httpx==0.28.1
requests==2.32.3